Move ETag computation to the interface level
These methods felt implementation-independent
